Organize directory according to MVVM architecture pattern
