--- unifood/lib/features/restaurants_filtered/widgets/search_app_bar.dart
@@ -1,92 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:unifood/widgets/custom_circled_button.dart';
-
-class SearchAppBar extends StatelessWidget implements PreferredSizeWidget {
-  final TextEditingController searchController;
-  final Function(String) onSearchChanged;
-  final VoidCallback onBackButtonPressed;
-
-  const SearchAppBar({
-    Key? key,
-    required this.searchController,
-    required this.onSearchChanged,
-    required this.onBackButtonPressed,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return AppBar(
-      automaticallyImplyLeading: false,
-      backgroundColor: Colors.transparent,
-      elevation: 0,
-      bottom: PreferredSize(
-        preferredSize: const Size.fromHeight(40.0),
-        child: Padding(
-          padding: const EdgeInsets.all(20),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                children: [
-                  CustomCircledButton(
-                    onPressed: onBackButtonPressed,
-                    diameter: 28,
-                    icon: const Icon(
-                      Icons.chevron_left_sharp,
-                      color: Colors.black,
-                    ),
-                    buttonColor: Colors.white,
-                  ),
-                  const SizedBox(width: 10),
-                  Expanded(
-                    child: Container(
-                      margin: const EdgeInsets.symmetric(horizontal: 2.5),
-                      decoration: BoxDecoration(
-                        color: Colors.white,
-                        borderRadius: BorderRadius.circular(20),
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.grey.withOpacity(0.5),
-                            spreadRadius: 2,
-                            blurRadius: 4,
-                            offset: const Offset(0, 3),
-                          ),
-                        ],
-                      ),
-                      child: TextField(
-                        controller: searchController,
-                        onChanged: onSearchChanged,
-                        decoration: InputDecoration(
-                          labelText: 'What do you want to eat?',
-                          labelStyle: const TextStyle(
-                            color: Color.fromARGB(255, 128, 126, 126),
-                            fontSize: 13,
-                          ),
-                          border: InputBorder.none,
-                          contentPadding:
-                              const EdgeInsets.only(left: 20, right: 20),
-                          suffixIcon: searchController.text.isNotEmpty
-                              ? IconButton(
-                                  icon: const Icon(Icons.clear),
-                                  onPressed: () {
-                                    searchController.clear();
-                                    onSearchChanged('');
-                                  },
-                                )
-                              : null,
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  @override
-  Size get preferredSize => const Size.fromHeight(kToolbarHeight + 40.0);
-}
