+++ unifood/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
 import 'package:flutter/material.dart';
-import 'package:unifood/routes.dart';
+import 'package:unifood/utils/routes.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
