+++ unifood/lib/view/restaurant/detail/widgets/details_bar.dart
