+++ unifood/lib/view/restaurant/search/widgets/restaurant_logo.dart
